+++ src/catalyst_sdwan_lab/tasks/deploy.py
@@ -142,6 +142,7 @@ def run(
                     proxy_ip=proxy_ip, proxy_port=proxy_port, no_proxy=no_proxy,
                 )
 
+                update("Adding control components...")
                 onboard_control_components(
                     client,
                     certs,